Add script to compute dependencies of many ports.
When building many top-level ports, many of their dependencies will be
held in common. We still need to build them all explicitly, however;
this script determines what the complete set of ports to build is.

We don't do anything frightfully clever like build the ports in an
optimal order: that's the kind of thing that Poudriere does. This rough
script does, however, give us the opportunity to get things building.